# modulo_administracion_vehiculos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Vehículos",
        "description": "Documentación de la API del sistema de administración de vehículos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/tipos-vehiculos": {
            "get": {
                "tags": [
                    "TipoVehiculo"
                ],
                "summary": "Listar todos los tipos de vehículos",
                "operationId": "ab2a7c33953d0246591f5198d2b5e52f",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de vehículos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TipoVehiculo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TipoVehiculo"
                ],
                "summary": "Crear un tipo de vehículo",
                "operationId": "aefb224d9164681aabe2e5c9047a01a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descripcion",
                                    "capacidad",
                                    "transmision"
                                ],
                                "properties": {
                                    "descripcion": {
                                        "type": "string",
                                        "example": "SUV"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "transmision": {
                                        "type": "string",
                                        "example": "Automática"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de vehículo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoVehiculo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tipos-vehiculos/{id}": {
            "get": {
                "tags": [
                    "TipoVehiculo"
                ],
                "summary": "Obtener un tipo de vehículo por ID",
                "operationId": "6a16883994293921e64877514f96ef37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de vehículo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoVehiculo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "TipoVehiculo"
                ],
                "summary": "Actualizar un tipo de vehículo",
                "operationId": "c01e3bd8d90fb1c58f7b30ac0f05f6ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Pickup"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "transmision": {
                                        "type": "string",
                                        "example": "Manual"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de vehículo actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoVehiculo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TipoVehiculo"
                ],
                "summary": "Eliminar un tipo de vehículo",
                "operationId": "318cedb235fa774afa27d9338fd40c87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado correctamente"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "TipoVehiculo": {
                "required": [
                    "descripción",
                    "capacidad",
                    "transmisión"
                ],
                "properties": {
                    "id_tipo": {
                        "type": "integer",
                        "example": 1
                    },
                    "descripción": {
                        "type": "string",
                        "example": "Camioneta"
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 5
                    },
                    "transmisión": {
                        "type": "string",
                        "example": "Manual"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}